# docs/33-openapi.yaml
openapi: 3.0.3
info:
  title: School Committee Platform API
  version: 0.1.0
  description: Frontend-facing BFF API contracts for MVP implementation.
servers:
  - url: https://platform.example.com/api
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      summary: List tasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: classId
          schema:
            type: string
      responses:
        "200":
          description: Task list
    post:
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
...
  /feedback:
    post:
      summary: Submit feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFeedbackRequest"
      responses:
        "201":
          description: Feedback submitted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateTaskRequest:
      type: object
      required:
        - title
        - description
        - priority
      properties:
        title:
          type: string
        description:
          type: string
        classId:
          type: string
          nullable: true
        deadline:
          type: string
          format: date
        priority:
          type: string
          enum: [low, normal, high, urgent]
        photoFileId:
          type: string
...
    SubmitFeedbackRequest:
      type: object
      required:
        - category
        - type
        - text
      properties:
        category:
          type: string
        type:
          type: string
          enum: [idea, complaint, suggestion, thanks]
        classId:
          type: string
        isAnonymous:
          type: boolean
        text:
          type: string